# core-common-distributor-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-distributor-mgmt
  description: Distributor Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/distributors
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/distributors/{distributorId}/simulations/{simulationId}:
    get:
      tags:
        - Distributor Simulations
      summary: Get distributor simulation by ID
      description: Returns a distributor simulation based on its ID
      operationId: getDistributorSimulationById
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: simulationId
          in: path
          description: ID of the distributor simulation
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved distributor simulation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '404':
          description: Distributor simulation not found
        '500':
          description: Internal server error
    put:
      tags:
        - Distributor Simulations
      summary: Update distributor simulation
      description: Updates an existing distributor simulation
      operationId: updateDistributorSimulation
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: simulationId
          in: path
          description: ID of the distributor simulation to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorSimulationDTO'
        required: true
      responses:
        '200':
          description: Distributor simulation successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '400':
          description: Invalid distributor simulation data provided
        '404':
          description: Distributor simulation not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Distributor Simulations
      summary: Delete distributor simulation
      description: Deletes a distributor simulation
      operationId: deleteDistributorSimulation
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: simulationId
          in: path
          description: ID of the distributor simulation to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Distributor simulation successfully deleted
        '404':
          description: Distributor simulation not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}:
    get:
      tags:
        - Distributor
      summary: Get distributor by ID
      description: Returns a distributor based on its ID
      operationId: getDistributorById
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved distributor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    put:
      tags:
        - Distributor
      summary: Update distributor
      description: Updates an existing distributor with the provided information
      operationId: updateDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorDTO'
        required: true
      responses:
        '200':
          description: Distributor successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorDTO'
        '400':
          description: Invalid distributor data provided
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Distributor
      summary: Delete distributor
      description: Deletes a distributor based on its ID
      operationId: deleteDistributor
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Distributor successfully deleted
        '404':
          description: Distributor not found
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/simulations/filter:
    post:
      tags:
        - Distributor Simulations
      summary: Filter distributor simulations
      description: Returns a paginated list of distributor simulations based on filter criteria
      operationId: filterDistributorSimulations
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorSimulationDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved distributor simulations
          content:
            application/json: {}
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
  /api/v1/distributors/{distributorId}/simulations:
    get:
      tags:
        - Distributor Simulations
      summary: Get all simulations for distributor
      description: Returns all simulations tracked by a distributor
      operationId: getSimulationsByDistributorId
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved distributor simulations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '500':
          description: Internal server error
    post:
      tags:
        - Distributor Simulations
      summary: Create a new distributor simulation
      description: Creates a new simulation tracking for a distributor
      operationId: createDistributorSimulation
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributorSimulationDTO'
        required: true
      responses:
        '201':
          description: Distributor simulation successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '400':
          description: Invalid distributor simulation data provided
        '409':
          description: Distributor simulation already exists for this application
        '500':
          description: Internal server error
...
  /api/v1/distributors/filter:
    post:
      tags:
        - Distributor
      summary: Filter distributors
      description: Returns a paginated list of distributors based on filter criteria
      operationId: filterDistributors
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestDistributorDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved distributors
          content:
            application/json: {}
        '400':
          description: Invalid filter criteria provided
        '500':
          description: Internal server error
...
  /api/v1/distributors/{distributorId}/simulations/{simulationId}/status:
    patch:
      tags:
        - Distributor Simulations
      summary: Update simulation status
      description: Updates the status of a distributor simulation
      operationId: updateSimulationStatus
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: simulationId
          in: path
          description: ID of the distributor simulation
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: New status for the simulation
          required: true
          schema:
            type: string
        - name: updatedBy
          in: query
          description: ID of the user performing the update
          required: false
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Simulation status successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '404':
          description: Distributor simulation not found
        '500':
          description: Internal server error
  /api/v1/distributors/{distributorId}/simulations/{simulationId}/deactivate:
    patch:
      tags:
        - Distributor Simulations
      summary: Deactivate distributor simulation
      description: Deactivates a distributor simulation
      operationId: deactivateDistributorSimulation
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: simulationId
          in: path
          description: ID of the distributor simulation to deactivate
          required: true
          schema:
            type: string
            format: uuid
        - name: updatedBy
          in: query
          description: ID of the user performing the update
          required: false
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Distributor simulation successfully deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '404':
          description: Distributor simulation not found
        '500':
          description: Internal server error
  /api/v1/distributors/{distributorId}/simulations/{simulationId}/activate:
    patch:
      tags:
        - Distributor Simulations
      summary: Activate distributor simulation
      description: Activates a distributor simulation
      operationId: activateDistributorSimulation
      parameters:
        - name: distributorId
          in: path
          description: ID of the distributor
          required: true
          schema:
            type: string
            format: uuid
        - name: simulationId
          in: path
          description: ID of the distributor simulation to activate
          required: true
          schema:
            type: string
            format: uuid
        - name: updatedBy
          in: query
          description: ID of the user performing the update
          required: false
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Distributor simulation successfully activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '404':
          description: Distributor simulation not found
        '500':
          description: Internal server error
...
  /api/v1/simulations/status/{status}:
    get:
      tags:
        - Simulations
      summary: Get simulations by status
      description: Returns all simulations with a specific status
      operationId: getSimulationsByStatus
      parameters:
        - name: status
          in: path
          description: Status of the simulations
          required: true
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved simulations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributorSimulationDTO'
        '500':
          description: Internal server error
...
components:
  schemas:
...
    DistributorSimulationDTO:
      required:
        - applicationId
        - distributorId
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        distributorId:
          type: string
          format: uuid
        applicationId:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
        agencyId:
          type: string
          format: uuid
        simulationStatus:
          maxLength: 50
          minLength: 0
          type: string
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: string
          format: uuid
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        updatedBy:
          type: string
          format: uuid
          readOnly: true
...
    FilterRequestDistributorSimulationDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/DistributorSimulationDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...